fix: harden Managed Controls field validation
