--- .github/workflows/managed-controls-batch13-push-remediation.yml
@@ -1,120 +1,0 @@
-name: Managed Controls batch 13 push remediation
-
-on:
-  push:
-    branches:
-      - managed-controls/batch-13-local-parser-shared-posture
-    paths:
-      - '.github/workflows/managed-controls-batch13-push-remediation.yml'
-      - 'scripts/ci/remediate_managed_controls_batch13.py'
-  workflow_dispatch:
-
-permissions:
-  contents: write
-
-concurrency:
-  group: managed-controls-batch13-push-remediation
-  cancel-in-progress: false
-
-jobs:
-  remediate:
-    if: github.event_name == 'workflow_dispatch' || !contains(github.event.head_commit.message, 'fix: harden Managed Controls field validation')
-    runs-on: ubuntu-24.04
-    timeout-minutes: 35
-    steps:
-      - name: Check out exact branch
-        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10
-        with:
-          ref: managed-controls/batch-13-local-parser-shared-posture
-          fetch-depth: 0
-
-      - name: Set up Python
-        uses: actions/setup-python@a309ff8b426b58ec0e2a45f0f869d46889d02405
-        with:
-          python-version: '3.12'
-
-      - name: Set up uv
-        uses: astral-sh/setup-uv@fac544c07dec837d0ccb6301d7b5580bf5edae39
-        with:
-          version: '0.9.26'
-          enable-cache: true
-
-      - name: Install dependencies
-        run: uv sync --frozen --extra dev --python 3.12
-
-      - name: Apply remediation
-        run: uv run --no-sync python scripts/ci/remediate_managed_controls_batch13.py
-
-      - name: Format and lint
-        run: |
-          uv run --no-sync ruff format \
-            src/codex_plugin_scanner/guard/managed_controls_policy_fields.py \
-            src/codex_plugin_scanner/guard/managed_controls_policy_bundle.py \
-            tests/test_managed_controls_policy_fields.py
-          uv run --no-sync ruff check --fix \
-            src/codex_plugin_scanner/guard/managed_controls_policy_fields.py \
-            src/codex_plugin_scanner/guard/managed_controls_policy_bundle.py \
-            tests/test_managed_controls_policy_fields.py
-          uv run --no-sync ruff format \
-            src/codex_plugin_scanner/guard/managed_controls_policy_fields.py \
-            src/codex_plugin_scanner/guard/managed_controls_policy_bundle.py \
-            tests/test_managed_controls_policy_fields.py
-
-      - name: Run focused verification
-        run: |
-          uv run --no-sync pytest -q \
-            tests/test_managed_controls_policy_fields.py \
-            tests/test_policy_bundle_v2.py \
-            --tb=short
-          uv run --no-sync basedpyright --level error \
-            src/codex_plugin_scanner/guard/managed_controls_policy_fields.py \
-            src/codex_plugin_scanner/guard/managed_controls_policy_bundle.py
-
-      - name: Refresh source-bound evidence
-        run: |
-          uv run --no-sync python tests/guard_command_decision_diff.py --write
-          uv run --no-sync python tests/guard_command_decision_diff.py --check
-
-      - name: Refresh reviewed test-suite ratchet
-        run: |
-          uv run --no-sync python scripts/ci/test_inventory.py --output test-inventory.json
-          uv run --no-sync python - <<'PY'
-          import json
-          from pathlib import Path
-
-          inventory = json.loads(Path('test-inventory.json').read_text(encoding='utf-8'))
-          baseline_path = Path('ci/test-suite-ratchet-baseline.json')
-          baseline = json.loads(baseline_path.read_text(encoding='utf-8'))
-          baseline['maximum_collected_cases'] = inventory['inventory']['collected_cases']
-          baseline['maximum_test_source_lines'] = inventory['test_source_lines']
-          baseline_path.write_text(
-              json.dumps(baseline, indent=2, sort_keys=True) + '\n',
-              encoding='utf-8',
-          )
-          PY
-          uv run --no-sync python scripts/ci/test_suite_ratchet.py \
-            --baseline ci/test-suite-ratchet-baseline.json \
-            --inventory test-inventory.json
-
-      - name: Enforce structural quality
-        run: |
-          uv run --no-sync python scripts/ci/code_quality_audit.py \
-            --root . \
-            --baseline ci/code-quality-baseline.json \
-            --json-output code-quality-audit.json
-          git diff --check
-
-      - name: Commit reviewed remediation
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          rm -f \
-            .github/workflows/managed-controls-batch13-push-remediation.yml \
-            .github/workflows/managed-controls-batch13-review-remediation.yml \
-            scripts/ci/remediate_managed_controls_batch13.py \
-            code-quality-audit.json \
-            test-inventory.json
-          git add -A
-          git diff --cached --check
-          git commit -m "fix: harden Managed Controls field validation"
-          git push origin HEAD:managed-controls/batch-13-local-parser-shared-posture
