+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
-  "maximum_collected_cases": 15326,
-  "maximum_test_source_lines": 330600,
+  "maximum_collected_cases": 15329,
+  "maximum_test_source_lines": 330674,
   "schema_version": 1
 }
